code refactoring

+++ controllers/movies.js
@@ -1,18 +1,14 @@
 const { Error: { ValidationError, CastError } } = require('mongoose');
 const Movie = require('../models/movie');
-const {
-  SUCCES_201,
-  NO_RIGHTS_TO_DELETE_ERROR_MESSAGE,
-  INCORRECT_DATA_MESSAGE,
-  MOVIE_NOT_FOUND_MESSAGE,
-} = require('../utils/constants');
 const BadRequestError = require('../errors/BadRequestError');
 const ForbiddenError = require('../errors/ForbiddenError');
 const NotFoundError = require('../errors/NotFoundError');
-
-const checkData = (data) => {
-  if (!data) throw new NotFoundError(MOVIE_NOT_FOUND_MESSAGE);
-};
+const {
+  HTTP_STATUS_CREATED,
+  NO_RIGHTS_TO_DELETE_MOVIE,
+  INCORRECT_MOVIE_DATA,
+  MOVIE_NOT_FOUND,
+} = require('../utils/constants');
 
 // Получение списка сохраненных пользователем фильмов
 module.exports.getMovies = (req, res, next) => {
@@ -23,46 +19,47 @@ module.exports.getMovies = (req, res, next) => {
     .then((movies) => res.send(movies))
     .catch((error) => {
       if (error instanceof ValidationError) {
-        return next(new BadRequestError(INCORRECT_DATA_MESSAGE));
+        return next(new BadRequestError(INCORRECT_MOVIE_DATA));
       }
       return next(error);
     });
 };
 
-// Создание нового фильма
+// Добавление фильма
 module.exports.addMovie = (req, res, next) => {
   const owner = req.user._id;
   const movieData = { ...req.body, owner };
 
   Movie.create(movieData)
     .then((movie) => movie.populate('owner'))
-    .then((populatedMovie) => res.status(SUCCES_201).send(populatedMovie))
+    .then((populatedMovie) => res.status(HTTP_STATUS_CREATED).send(populatedMovie))
     .catch((error) => {
       if (error instanceof ValidationError) {
-        return next(new BadRequestError(INCORRECT_DATA_MESSAGE));
+        return next(new BadRequestError(INCORRECT_MOVIE_DATA));
       }
       return next(error);
     });
 };
 
+// Удаление фильма
 module.exports.deleteMovie = (req, res, next) => {
-  const movieId = req.params;
-
+  const { _id: movieId } = req.params;
   Movie.findById(movieId)
     .then((movie) => {
-      checkData(movie);
+      if (!movie) throw new NotFoundError(MOVIE_NOT_FOUND);
 
       const ownerId = movie.owner.valueOf();
       const userId = req.user._id;
+
       if (ownerId !== userId) {
-        return Promise.reject(new ForbiddenError(NO_RIGHTS_TO_DELETE_ERROR_MESSAGE));
+        return next(new ForbiddenError(NO_RIGHTS_TO_DELETE_MOVIE));
       }
       return movie.deleteOne();
     })
     .then((deletedMovie) => res.send(deletedMovie))
     .catch((error) => {
       if (error instanceof CastError) {
-        return next(new BadRequestError(INCORRECT_DATA_MESSAGE));
+        return next(new BadRequestError(INCORRECT_MOVIE_DATA));
       }
       return next(error);
     });
